c5: lexer token is now a Tok newtype, no more `as i64` casts
`Lexer::tk` was `i64`, so every comparison site read
`self.lex.tk == Token::MulOp as i64`. The cast was load-bearing
only because c5 mixes ASCII punctuation (raw byte values) and
multi-character keywords / operators (the `Token` enum starting
at 128) in the same scalar slot.

Wrap the field in `pub struct Tok(pub i64)` and implement
`PartialEq<Token>`, `PartialEq<char>`, `PartialEq<u8>`,
`PartialEq<i32>`, `PartialEq<i64>` (plus `PartialOrd<i64>` for
range tests and `Display` for diagnostics) so all four shapes
`tk == Token::X`, `tk == '('`, `tk == b';'`, `tk == 0` work
without the cast. `Tok::raw()` exposes the underlying integer
for the few sites that genuinely need the bit pattern
(`token::describe`, the test harness's `next() -> i64`,
ordering checks).

Pure refactor; all existing tests + stb smoke green.

Closes #83.

+++ src/c5/compiler/aggregate.rs
@@ -91,7 +91,7 @@ impl Compiler {
         let mut bf_active = false;
         let mut bf_storage_offset: usize = 0;
         let mut bf_next_bit: u32 = 0;
-        while self.lex.tk != '}' as i64 {
+        while self.lex.tk != '}' {
             // Field type prefix: int, char, float, double, or struct Name.
             // Leading qualifiers / int modifiers / function specifiers
             // (`const`, `unsigned`, ...) are no-ops; track if any int
@@ -112,18 +112,18 @@ impl Compiler {
             // synthesised tag stays a regular nested-struct type.
             let mut anon_aggregate_inner_id: Option<usize> = None;
             while is_decl_modifier(self.lex.tk) {
-                if self.lex.tk == Token::IntMod as i64 {
+                if self.lex.tk == Token::IntMod {
                     saw_int_mod = true;
-                } else if self.lex.tk == Token::Signed as i64 {
+                } else if self.lex.tk == Token::Signed {
                     saw_signed = true;
                     saw_int_mod = true;
-                } else if self.lex.tk == Token::Unsigned as i64 {
+                } else if self.lex.tk == Token::Unsigned {
                     saw_unsigned = true;
                     saw_int_mod = true;
-                } else if self.lex.tk == Token::Long as i64 {
+                } else if self.lex.tk == Token::Long {
                     long_count = long_count.saturating_add(1);
                     saw_int_mod = true;
-                } else if self.lex.tk == Token::Short as i64 {
+                } else if self.lex.tk == Token::Short {
                     saw_short = true;
                     saw_int_mod = true;
                 }
@@ -131,7 +131,7 @@ impl Compiler {
             }
             let saw_long = long_count >= 1;
             let saw_long_long = long_count >= 2;
-            let field_base = if self.lex.tk == Token::Int as i64 {
+            let field_base = if self.lex.tk == Token::Int {
                 self.next()?;
                 let base = if saw_long_long {
                     Ty::LongLong as i64
@@ -147,7 +147,7 @@ impl Compiler {
                 } else {
                     base
                 }
-            } else if self.lex.tk == Token::Char as i64 {
+            } else if self.lex.tk == Token::Char {
                 self.next()?;
                 // Mirror parse_decl_base_type: `signed char` is a
                 // real 1-byte signed field; plain / unsigned char
@@ -157,14 +157,14 @@ impl Compiler {
                 } else {
                     Ty::Char as i64 | UNSIGNED_BIT
                 }
-            } else if self.lex.tk == Token::Float as i64 {
+            } else if self.lex.tk == Token::Float {
                 self.next()?;
                 Ty::Float as i64
-            } else if self.lex.tk == Token::Double as i64 {
+            } else if self.lex.tk == Token::Double {
                 self.next()?;
                 Ty::Double as i64
-            } else if self.lex.tk == Token::Struct as i64 || self.lex.tk == Token::Union as i64 {
-                let nested_is_union = self.lex.tk == Token::Union as i64;
+            } else if self.lex.tk == Token::Struct || self.lex.tk == Token::Union {
+                let nested_is_union = self.lex.tk == Token::Union;
                 self.next()?;
                 // Three shapes:
                 //   * `struct Foo { ... }` -- named definition.
@@ -180,11 +180,11 @@ impl Compiler {
                 //                            fields PROMOTE into
                 //                            the enclosing scope
                 //                            -- C11 6.7.2.1p13.
-                let (inner_name, had_explicit_tag) = if self.lex.tk == Token::Id as i64 {
+                let (inner_name, had_explicit_tag) = if self.lex.tk == Token::Id {
                     let name = self.symbols[self.lex.curr_id_idx].name.clone();
                     self.next()?;
                     (name, true)
-                } else if self.lex.tk == '{' as i64 {
+                } else if self.lex.tk == '{' {
                     let kind = if nested_is_union {
                         "anon_union"
                     } else {
@@ -197,7 +197,7 @@ impl Compiler {
                 } else {
                     return Err(self.compile_err("aggregate name or `{{` expected in field type"));
                 };
-                let inner_id = if self.lex.tk == '{' as i64 {
+                let inner_id = if self.lex.tk == '{' {
                     self.parse_aggregate_body(&inner_name, nested_is_union)?
                 } else {
                     self.find_or_forward_declare_struct(&inner_name)
@@ -208,7 +208,7 @@ impl Compiler {
                     Some(inner_id)
                 };
                 struct_ty_for(inner_id)
-            } else if self.lex.tk == Token::Enum as i64 {
+            } else if self.lex.tk == Token::Enum {
                 // C99 6.7.2.2: an `enum X` field collapses to plain
                 // `int` in c5's type system the same way every other
                 // enum reference does. Consume any tag name and the
@@ -217,10 +217,10 @@ impl Compiler {
                 // `parse_decl_base_type` enum branch so the same
                 // shape works at file scope and inside a struct.
                 self.next()?;
-                if self.lex.tk == Token::Id as i64 {
+                if self.lex.tk == Token::Id {
                     self.next()?;
                 }
-                if self.lex.tk == '{' as i64 {
+                if self.lex.tk == '{' {
                     self.parse_enum_body()?;
                 }
                 Ty::Int as i64
@@ -264,7 +264,7 @@ impl Compiler {
             // an unnamed `struct { ... }` and must be reachable
             // as `li.LowPart` (not `li.<some-tag>.LowPart`).
             if let Some(inner_id) = anon_aggregate_inner_id
-                && self.lex.tk == ';' as i64
+                && self.lex.tk == ';'
             {
                 {
                     // Seal any pending bitfield run -- the
@@ -333,7 +333,7 @@ impl Compiler {
                 // Anonymous bitfield (`int :N;`) -- skips a name and
                 // just reserves bits for padding. Detected by `:`
                 // appearing in declarator position.
-                let anon_bitfield_width = if self.lex.tk == ':' as i64 {
+                let anon_bitfield_width = if self.lex.tk == ':' {
                     self.next()?;
                     let n = self.parse_constant_int()?;
                     if n < 0 {
@@ -364,7 +364,7 @@ impl Compiler {
                         }
                         bf_next_bit += width;
                     }
-                    if self.lex.tk == ',' as i64 {
+                    if self.lex.tk == ',' {
                         self.next()?;
                         continue;
                     }
@@ -399,7 +399,7 @@ impl Compiler {
                 let mut bit_width: u32 = 0;
                 let mut bit_offset: u32 = 0;
                 let field_offset: usize;
-                if self.lex.tk == ':' as i64 {
+                if self.lex.tk == ':' {
                     if field_array_size != 0 {
                         return Err(self.compile_err("array fields cannot also be bitfields"));
                     }
@@ -489,14 +489,14 @@ impl Compiler {
                     bit_width,
                 });
 
-                if self.lex.tk == ',' as i64 {
+                if self.lex.tk == ',' {
                     self.next()?;
                     continue;
                 }
                 break;
             }
 
-            if self.lex.tk != ';' as i64 {
+            if self.lex.tk != ';' {
                 return Err(self.compile_err("semicolon expected after struct field"));
             }
             self.next()?;
